Die Auflösung der Bilder scheint nicht mehr zu funktionieren - habs in Postman getestet und bekomme immer die gleich große Antwort zurück 	Bei createArticle akzeptiert der keine 13 stelligen EANs 	Article löschen funktioniert garnicht (immer nen 417) 	die E-Mails scheinen beim Anmelden casesensitiv sein

+++ src/main/java/de/fhdw/models/Article.java
@@ -19,7 +19,7 @@ public class Article extends PanacheEntityBase {
     @Column(nullable = false)
     public double price;
 
-    public int ean;
+    public double ean;
 
     public String description;
